LL2/requirements.txt
sourcesourcrewai==0.28.8
crewai-tools==0.1.6
langchain>=0.1.10,<0.2.0
langchain-openai==0.0.5
python-dotenv==1.0.0
pydantic>=2.6.1,<3.0.0
PyYAML==6.0.1
